apps/rag-task-processor/pyproject.toml
[project]
name = "rag-task-processor"
version = "0.1.0"
description = "Cloud Function that processes Cloud Tasks and executes Cloud Run Jobs for document processing"
requires-python = ">=3.10"
dependencies = [
    "functions-framework>=3.4.0",
    "google-cloud-run>=0.10.0",
    "structlog>=23.2.0",
    "grpcio>=1.60.0,<1.75.0",
    "flask>=3.1.0",
]

# Cloud Functions optimization - prevent source builds
[tool.uv]
no-build-package = ["grpcio"]  # Don't build grpcio from source - use wheels only

[dependency-groups]
dev = [
    "mypy>=1.6.0",
    "types-protobuf>=4.24.0",
]
lint = [
    "ruff>=0.1.0",
    "black>=23.0.0", 
    "mypy>=1.6.0",
]

